Implementação semi-final (sem GUI)

+++ src/crescente.h
@@ -6,6 +6,6 @@
 
 #define NBARALHOS 2
 
-int jogar();
+unsigned long jogar();
 
 int main();
